Implement rest of LuaInstanceBridge
matches aren't very good but are as best as they can be(?) for now

+++ Client/App/include/script/ThreadRef.h
@@ -92,7 +92,7 @@ namespace RBX
 
         public:
             FunctionRef(const FunctionRef&);
-            FunctionRef(lua_State*, size_t);
+            FunctionRef(lua_State*, int);
             FunctionRef();
             ~FunctionRef();
